Move `_validate_color_sequence` into `pyvista.plotting.colors`
The helper builds `Color` objects, so it belongs with them rather than in
the validation package, which is being extracted.
