+++ pyvista/core/_validation/validate.py
@@ -41,8 +41,6 @@ from pyvista.core._validation._cast_array import _cast_to_numpy
 from pyvista.core._validation._cast_array import _cast_to_tuple
 
 if TYPE_CHECKING:
-    from collections.abc import Sequence
-
     import numpy.typing as npt
 
     from pyvista.core._typing_core import ArrayLike
@@ -51,8 +49,6 @@ if TYPE_CHECKING:
     from pyvista.core._typing_core import TransformLike
     from pyvista.core._typing_core import VectorLike
     from pyvista.core._typing_core._array_like import NumpyArray
-    from pyvista.plotting._typing import ColorLike
-    from pyvista.plotting.colors import Color
 
     from .check import _ShapeLike
 
@@ -1262,46 +1258,3 @@ def validate_dimensionality(
     _set_default_kwarg_mandatory(kwargs, 'must_have_shape', shape)
 
     return validate_array(dimensionality_as_array, **kwargs)
-
-
-def _validate_color_sequence(
-    color: ColorLike | Sequence[ColorLike],
-    n_colors: int | None = None,
-) -> tuple[Color, ...]:
-    """Validate a color sequence.
-
-    If ``n_colors`` is specified, the output will have ``n`` colors. For single-color
-    inputs, the color is copied and a sequence of ``n`` identical colors is returned.
-    For inputs with multiple colors, the number of colors in the input must
-    match ``n_colors``.
-
-    If ``n_colors`` is None, no broadcasting or length-checking is performed.
-    """
-    from pyvista.plotting.colors import Color  # noqa: PLC0415
-
-    try:
-        # Assume we have one color
-        color_list = [Color(color)]
-        n_colors = 1 if n_colors is None else n_colors
-        return tuple(color_list * n_colors)
-    except ValueError:
-        if isinstance(color, (tuple, list)):
-            try:
-                color_list = [_validate_color_sequence(c, n_colors=1)[0] for c in color]
-                if len(color_list) == 1:
-                    n_colors = 1 if n_colors is None else n_colors
-                    color_list = color_list * n_colors
-
-                # Only return if we have the correct number of colors
-                if n_colors is None or len(color_list) == n_colors:
-                    return tuple(color_list)
-            except ValueError:
-                pass
-    n_colors_str = f' {n_colors} ' if n_colors else ' '
-    msg = (
-        f'Invalid color(s):\n'
-        f'\t{color}\n'
-        f'Input must be a single ColorLike color '
-        f'or a sequence of{n_colors_str}ColorLike colors.'
-    )
-    raise ValueError(msg)
